pyproject.toml
[project]
name = "cdumm"
version = "4.0.0"
description = "Crimson Desert Elite BR — gerenciador de mods para Crimson Desert, by CrashByte"
requires-python = ">=3.10"
dependencies = [
    "pyside6-essentials>=6.6",
    "PySide6-Fluent-Widgets>=1.11",  # used by bug_report.py (dialog sob demanda)
    "bsdiff4>=1.2",
    "lz4>=4.3",
    "cryptography>=42.0",
    "xxhash>=3.0",                   # apply_engine.py, snapshot_manager.py
    "py7zr>=0.20",                   # import_handler.py, asi_manager.py, main_window.py
    "lxml>=5.0",                     # xml_patch_handler.py (engine Faisal)
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pyinstaller>=6.0",
    "ruff>=0.4",
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
